# Optimize filling a region with a solid brush when antialias is off