--- a/src/ImageSharp.Drawing/Processing/Processors/Drawing/FillRegionProcessor.cs
+++ b/src/ImageSharp.Drawing/Processing/Processors/Drawing/FillRegionProcessor.cs
@@ -3,7 +3,7 @@
 
 using System;
 using System.Buffers;
-
+using SixLabors.ImageSharp.Advanced;
 using SixLabors.ImageSharp.Memory;
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Primitives;
@@ -168,16 +168,30 @@ namespace SixLabors.ImageSharp.Processing.Processors.Drawing
                         {
                             if (!this.Options.Antialias)
                             {
+                                bool hasOnes = false;
+                                bool hasZeros = false;
                                 for (int x = 0; x < scanlineWidth; x++)
                                 {
                                     if (scanline[x] >= 0.5)
                                     {
                                         scanline[x] = 1;
+                                        hasOnes = true;
                                     }
                                     else
                                     {
                                         scanline[x] = 0;
+                                        hasZeros = true;
                                     }
+                                }
+
+                                if (hasOnes != hasZeros && this.IsSolidBrushWithoutBlending(out SolidBrush<TPixel> solidBrush))
+                                {
+                                    if (hasOnes)
+                                    {
+                                        source.GetPixelRowSpan(y).Slice(minX, scanlineWidth).Fill(solidBrush.Color);
+                                    }
+
+                                    continue;
                                 }
                             }
 
@@ -186,6 +200,18 @@ namespace SixLabors.ImageSharp.Processing.Processors.Drawing
                     }
                 }
             }
+        }
+
+        private bool IsSolidBrushWithoutBlending(out SolidBrush<TPixel> solidBrush)
+        {
+            solidBrush = this.Brush as SolidBrush<TPixel>;
+
+            if (solidBrush == null)
+            {
+                return false;
+            }
+
+            return this.Options.IsOpaqueColorWithoutBlending(solidBrush.Color);
         }
     }
 }